Update Dockerfile
Update Dockerfile for compatibility with Girder Worker.

- Run as root instead of as core3d user.

- Update conda base to avoid errors when installing the environment [1].

When passing files to the container using GirderFileIdToVolume, Girder
Worker creates a temporary directory on the host and shares it with the
container. Because the temporary directory is created with owner-only
read permissions [2], only the root user can read the directory in the
container.

[1] https://github.com/conda/conda/issues/6811

[2] https://github.com/girder/girder_worker/blob/b21cd554332b3f34c9e268fe42f3882523dcc084/girder_worker/docker/transforms/__init__.py#L131