Require IMD file as command line argument
Remove the AOI parameter and require the caller of the
material-classifier.py tool to additionally pass the path to the IMD
file that corresponds to the input image.

Since the AOI parameter and the logic to find IMD files based on an
image file name is removed, this commit also removes the ability to run
the tool on a folder. Batch operations at that level should now be
driven outside of the tool.