Add wrapper tool to run Purdue roof geon extraction
Add a wrapper script to run the steps of Purdue roof geon extraction,
including segmentation and reconstruction. The step to convert from PLY
to OBJ is omitted, because that conversion tool is not yet available.

To support running through Girder Worker, where it's necessary to
specify an explicit output directory to be able to upload files back to
Girder, the script requires an output directory as a parameter.