Cherry-picked / applied changes from previous premerge branch
Prior to the Purdue x Columbia update, we had made some changes to the
roof_segmentation.py script in the
"columbia_geon_classification_premerge_fixes" branch.  This commit
applies the changes made to that branch on top of the latest code from
Columbia as a part of that Purdue x Columbia update.