Added arguments to segment_by_height tool to ease girder integration
When using the --road-vector argument, the script ingests several
files we only specify the full path to one of them.  This is
challenging to integrate in the app due to how Girder and Docker
coordinate files.  Splitting the path and prefix up in this case makes
the interaction a bit cleaner.