Added functionality to resume a failed run
Added a wrapper function for each of the pipeline steps called
'run_step'.  This function writes stdout and stderr to a log file, and
creates a file which records the exit status of the step.  When
'run_step' is called, if a successful exit status file exists for the
step already, the step is skipped.